refactor(canvas): share recorded wheel sample type
